feat: add support for LF to CRLF conversion warnings
This commit adds git add warnings for files that will have their line endings
converted from LF to CRLF when touched by Git. This is to provide developers
with more visibility into the line ending changes that will occur.

This change should help developers be more aware of line ending changes in the
repository and take appropriate action if needed.